Gzyl edit 18OCT20 • Coach / Swimmer Profiles work properly. Updates and additions to both profiles will work. • Adjusted tuition (changed to decimal instead of int) • Adjusted some of the links between VIews

+++ Models/Lesson.cs
@@ -11,6 +11,6 @@ namespace CourseProject.Models
         public int LessonId { get; set; }
         public string LessonTitle { get; set; }
         public string SkillLevel { get; set; }
-        public int Tuition { get; set; }
+        public decimal Tuition { get; set; }
     }
 }
